base: table drop order fix
* Fixes order in which tables are being dropped.
