+++ invenio/modules/editor/models.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of Invenio.
-# Copyright (C) 2011, 2012, 2013, 2014 CERN.
+# Copyright (C) 2011, 2012, 2013, 2014, 2015 CERN.
 #
 # Invenio is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License as
@@ -19,13 +19,10 @@
 
 """Editor database models."""
 
-# General imports.
 from invenio.ext.sqlalchemy import db
-
-# Create your models here.
-
 from invenio.modules.records.models import Record as Bibrec
 
+from sqlalchemy import event
 
 class BibHOLDINGPEN(db.Model):
 
@@ -166,6 +163,22 @@ class BibEDITCACHE(db.Model):
     is_active = db.Column(db.TinyInteger(1, unsigned=True),
                           server_default='1', nullable=False)
 
+
+def bibdoc_before_drop(target, connection_dummy, **kw_dummy):
+    import os
+    import shutil
+    from invenio.legacy.bibdocfile.api import _make_base_dir
+    print
+    print(">>> Going to remove records data...")
+    for (docid,) in db.session.query(target.c.id).all():
+        directory = _make_base_dir(docid)
+        if os.path.isdir(directory):
+            print('    >>> Removing files for docid =', docid)
+            shutil.rmtree(directory)
+    db.session.commit()
+    print(">>> Data has been removed.")
+
+event.listen(Bibdoc.__table__, "before_drop", bibdoc_before_drop)
 
 __all__ = ('Bibrec',
            'BibEDITCACHE',
